requirements.txt
streamlit==1.28.1
openai==1.3.0
langchain==0.0.350
beautifulsoup4==4.12.2
requests==2.31.0
google-search-results==2.4.2
kaggle==1.5.16
huggingface-hub==0.19.4
PyGithub==1.59.1
pandas==2.1.3
python-dotenv==1.0.0